Update navigation.blade.php
Revert changes from Pull Request #4 because it caused the tool not to work properly without Gates/Policies. It only hides the tool and does not protect the routes/resources fully. Fixes #9 and resolves the need to add `vyuldashev/nova-permission` for the Gates/Policies #18. Will think about hiding Resources in the future, but added a note to the readme to hide the tool with the `canSee` method.